requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
ConfigArgParse==1.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-filter==22.1
django-redis==5.2.0
django-rest-swagger==2.2.0
django-silk==5.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
drf-nested-routers==0.93.4
exceptiongroup==1.0.4
filelock==3.8.0
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
flower==1.2.0
gevent==22.10.2
geventhttpclient==2.0.8
gprof2dot==2022.7.29
greenlet==2.0.1
gunicorn==20.1.0
humanize==4.4.0
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
locust==2.13.0
MarkupSafe==2.1.1
model-bakery==1.9.0
msgpack==1.0.4
mysqlclient==2.1.1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.3
Pillow==9.3.0
pipenv==2022.11.11
platformdirs==2.5.4
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.4
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytest-watch==4.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
pyzmq==24.0.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
roundrobin==0.0.4
simplejson==3.18.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.7
virtualenv-clone==0.5.7
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.5.2
